api/requirements.txt
Flask==3.0.1
psycopg2-binary==2.9.5
requests
newsapi-python==0.2.6
redis
pandas
sqlalchemy
newspaper3K
feedparser
lxml_html_clean
python-dotenv
pgvector
hf_xet
sentence-transformers
pipeline